Move registryHandler into its own file into the image package
